requirements.txt
Red-DiscordBot>=3.5.0
discord.py>=2.0.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0
cryptography>=41.0.0
pydantic>=2.0.0
httpx>=0.24.0
tenacity>=8.2.0
